Create Onyibe chioma 2024714038
The python program implements a grading system using "If" and "elif" statement. It takes a student's score as input and print the corresponding grade based on the following scale:
70-100: Grade A
60-69: Grade B
50-59: Grade C
45-49 Grade D
0-39 Grade F
For each valid score, the program also print "yes" to confirm the grading was successful. If the valid range, it prints "invalid score" and "No".
Created by: Onyibe chioma 
Reg No: 2024714038